Rename : Basic 컴포넌트 index.js 파일 및 import 수정
