+++ src/Basic/Avatar/Avatar.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Avatar from "./Avatar";
+import Avatar from "./";
 
 export default {
   title: "Basic/Avatar",
