+++ src/Basic/Avatar/index.js
@@ -1,2 +1,1 @@
-export { default as Avatar } from "./Avatar";
-export { default as UserAlt } from "./UserAlt";
+export { default } from "./Avatar";
