+++ src/Basic/Icon/index.js
@@ -1,3 +1,1 @@
-export { default as Icon } from "./Icon";
-export { default as IconBookMark } from "./IconBookMark";
-export { default as IconSocial } from "./IconSocial";
+export { default } from "./Icon";
