+++ src/Basic/IconSocial/IconSocial.stories.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
 
-import Icon from "./";
-import { icons } from "../../shared/icons";
+import IconSocial from ".";
+import { iconsSocial } from "../../shared/iconsSocial";
 
 export default {
   title: "Basic/Icon",
@@ -34,28 +34,24 @@ const List = styled.ul`
   list-style: none;
 `;
 
-export const Default = (args) => (
+export const Social = () => (
   <>
-    This is a {args.block ? "block" : "inline"} <Icon {...args} /> icon
-  </>
-);
-
-Default.args = {
-  icon: "smile",
-  "aria-label": "smile",
-  block: false,
-};
-
-export const All = () => (
-  <>
-    There are {Object.keys(icons).length} icons
+    There are {Object.keys(iconsSocial).length} social icons
     <List>
-      {Object.keys(icons).map((key) => (
+      {Object.keys(iconsSocial).map((key) => (
         <Item key={key}>
-          <Icon icon={key} aria-hidden />
+          <IconSocial icon={key} aria-hidden />
           <Meta>{key}</Meta>
         </Item>
       ))}
     </List>
+  </>
+);
+
+export const BookMark = () => (
+  <>
+    <IconBookMark mode="primary" theme="light" />
+    <IconBookMark mode="secondary" theme="light" />
+    <IconBookMark mode="secondary" theme="dark" />
   </>
 );
