+++ src/Basic/Logo/index.js
@@ -1,1 +1,1 @@
-export { default as Logo } from "./Logo";
+export { default } from "./Logo";
